Touched up page rendering and changed database error reporting for async edits/deletions.

+++ VolunteerSystem/Controllers/VolunteersController.cs
@@ -60,13 +60,24 @@ namespace VolunteerSystem.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ID,FirstName,LastName,UserName,Password,Address,HomePhone,WorkPhone,CellPhone,Email,LicenseOnFile,SSCardOnFile,Approved")] Volunteer volunteer)
+        public async Task<IActionResult> Create([Bind("FirstName,LastName,UserName,Password,Address,HomePhone,WorkPhone,CellPhone,Email,LicenseOnFile,SSCardOnFile,Approved")] Volunteer volunteer)
         {
-            if (ModelState.IsValid)
+            try {
+
+                if (ModelState.IsValid)
+                {
+                    _context.Add(volunteer);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction("Index");
+                }
+            } 
+            catch (DbUpdateException /* ex */) 
             {
-                _context.Add(volunteer);
-                await _context.SaveChangesAsync();
-                return RedirectToAction("Index");
+                //Log the error (uncomment ex variable name and write a log.
+                ModelState.AddModelError("", "Unable to save changes. " +
+                "Try again, and if the problem persists " +
+                "see your system administrator.");
+
             }
             return View(volunteer);
         }
@@ -99,31 +110,34 @@ namespace VolunteerSystem.Controllers
                 return NotFound();
             }
 
-            if (ModelState.IsValid)
-            {
-                try
+            var volunteerToUpdate = await _context.Volunteers.SingleOrDefaultAsync(v => v.ID == id);
+
+            if (await TryUpdateModelAsync<Volunteer>(
+                volunteerToUpdate, 
+                "",
+                v => volunteer.FirstName, v => volunteer.LastName, v => volunteer.UserName, v => volunteer.Password, 
+                v => volunteer.Address, v => volunteer.HomePhone, v => volunteer.WorkPhone, v => volunteer.CellPhone,
+                v => volunteer.Email, v => volunteer.LicenseOnFile, v => volunteer.SSCardOnFile, v => volunteer.Approved))
                 {
-                    _context.Update(volunteer);
+                try 
+                {
                     await _context.SaveChangesAsync();
+                    return RedirectToAction("Index");
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException /* ex */) 
                 {
-                    if (!VolunteerExists(volunteer.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                  //Log the error (uncomment ex variable name and write a log.)
+                    ModelState.AddModelError("", "Unable to save changes. " +
+                    "Try again, and if the problem persists, " +
+                    "see your system administrator.");
                 }
-                return RedirectToAction("Index");
             }
             return View(volunteer);
         }
+        
 
         // GET: Volunteers/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -131,11 +145,19 @@ namespace VolunteerSystem.Controllers
             }
 
             var volunteer = await _context.Volunteers
+                .AsNoTracking()
                 .SingleOrDefaultAsync(m => m.ID == id);
             if (volunteer == null)
             {
                 return NotFound();
             }
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                "Delete failed. Try again, and if the problem persists " +
+                "see your system administrator.";
+            }
+
 
             return View(volunteer);
         }
@@ -143,12 +165,30 @@ namespace VolunteerSystem.Controllers
         // POST: Volunteers/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(int? id, bool? saveChangesError = false)
         {
-            var volunteer = await _context.Volunteers.SingleOrDefaultAsync(m => m.ID == id);
-            _context.Volunteers.Remove(volunteer);
-            await _context.SaveChangesAsync();
-            return RedirectToAction("Index");
+            var volunteer = await _context.Volunteers
+                .AsNoTracking()
+                .SingleOrDefaultAsync(m => m.ID == id);
+            
+            if (volunteer == null) 
+            {
+                return RedirectToAction("Index");
+            }
+
+            try 
+            {
+                _context.Volunteers.Remove(volunteer);
+                await _context.SaveChangesAsync();
+                return RedirectToAction("Index");
+
+            } 
+            catch (DbUpdateException /* ex */) 
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                return RedirectToAction("Delete", new { id = id, saveChangesError = true });
+
+            }
         }
 
         private bool VolunteerExists(int id)
